backend/websocket-server/requirements.txt
Django==4.2.5
channels==4.0.0
channels-redis==4.1.0
redis==4.6.0
django-cors-headers==4.3.1
djangorestframework==3.14.0
kafka-python==2.0.2